core/standards/css3/image-fit/js/ua-stylesheet/object/001.html

<!doctype html>
<title>object default style</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<object></object>
<script>
test(function() {
  var e = document.body.firstChild;
  assert_equals(getComputedStyle(e, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
});
</script>
